Vim: Fix LSP for mason-lspconfig 2+ and Neovim 0.11+
This reverts previous commit "Vim: Disable Pyright" that wasn't
effective, and fixes the `servers = { … }` configuration that was never
really applied, because of a recent change in Mason.

I debugged it and found that it was caused by this change:
https://github.com/mason-org/mason-lspconfig.nvim/commit/80f2cd773
which affects other users:
https://github.com/nvim-lua/kickstart.nvim/issues/1595
and has ongoing pull requests to fix. To clean my homemade fix, I
inspired from this one:
https://github.com/nvim-lua/kickstart.nvim/pull/1590/files